+++ com.bistro.sagg.core/src/main/java/com/bistro/sagg/core/builders/ProductCategoryBuilder.java
@@ -1,0 +1,27 @@
+package com.bistro.sagg.core.builders;
+
+import com.bistro.sagg.core.model.products.ProductCategory;
+
+public class ProductCategoryBuilder {
+
+	private ProductCategory category;
+
+	public ProductCategoryBuilder() {
+		super();
+		this.category = new ProductCategory();
+	}
+
+	public ProductCategoryBuilder(ProductCategory category) {
+		super();
+		this.category = category;
+	}
+
+	public void build(String name) {
+		this.category.setName(name);
+	}
+
+	public ProductCategory getCategory() {
+		return category;
+	}
+
+}