+++ com.bistro.sagg.core/src/main/java/com/bistro/sagg/core/model/Product.java
@@ -1,5 +1,0 @@
-package com.bistro.sagg.core.model;
-
-public class Product {
-
-}