+++ com.bistro.sagg.core/src/main/java/com/bistro/sagg/core/repository/MarketableProductRepository.java
@@ -1,0 +1,9 @@
+package com.bistro.sagg.core.repository;
+
+import org.springframework.data.repository.CrudRepository;
+
+import com.bistro.sagg.core.model.products.MarketableProduct;
+
+public interface MarketableProductRepository extends CrudRepository<MarketableProduct, Long> {
+
+}