+++ com.bistro.sagg.core/src/main/java/com/bistro/sagg/core/repository/SupplyRepository.java
@@ -1,0 +1,9 @@
+package com.bistro.sagg.core.repository;
+
+import org.springframework.data.repository.CrudRepository;
+
+import com.bistro.sagg.core.model.products.Supply;
+
+public interface SupplyRepository extends CrudRepository<Supply, Long> {
+
+}